+++ src/main/java/com/lowdragmc/mbd2/common/graphprocessor/MachineEventGraphView.java
@@ -2,7 +2,6 @@ package com.lowdragmc.mbd2.common.graphprocessor;
 
 import com.lowdragmc.lowdraglib.gui.graphprocessor.data.BaseGraph;
 import com.lowdragmc.lowdraglib.gui.graphprocessor.widget.GraphViewWidget;
-import com.lowdragmc.mbd2.MBD2;
 
 import java.util.List;
 
@@ -15,11 +14,5 @@ public class MachineEventGraphView extends GraphViewWidget {
     protected void setupNodeGroups(List<String> supportNodeGroups) {
         super.setupNodeGroups(supportNodeGroups);
         supportNodeGroups.add("graph_processor.node.mbd2.machine");
-        if (MBD2.isGeckolibLoaded()) {
-            supportNodeGroups.add("graph_processor.node.mbd2.machine.geckolib");
-        }
-        if (MBD2.isPhotonLoaded()) {
-            supportNodeGroups.add("graph_processor.node.mbd2.machine.photon");
-        }
     }
 }