+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # ChangeLog
 
 * add item durability recipe capability.
+* add create rpm recipe capability.
+* remove duplicated nodes.